REF:moved execution nodes from module files"
execution node separation (modules cleaned up, nodes moved to execution/)
