+++ src/adapt/execution/nodes/__init__.py
@@ -1,3 +1,2 @@
 # Copyright © 2026, UChicago Argonne, LLC
 # See LICENSE for terms and disclaimer.
-
