+++ src/adapt/modules/base.py
@@ -1,50 +1,44 @@
 # Copyright © 2026, UChicago Argonne, LLC
 # See LICENSE for terms and disclaimer.
 
-"""Base interface for all Adapt processing modules.
+"""Base interface for all Adapt execution nodes.
 
-Every module in the system — whether in modules/ or extensions/ — must
-declare its name, inputs, and outputs. The graph engine uses these
-declarations to build the execution DAG automatically.
+Every node in the system — whether in execution/nodes/ or extensions/ — must
+declare its name, inputs, outputs, and optionally input/output contracts.
+The graph engine uses these declarations to build the execution DAG automatically.
 
-Existing scientific classes (AwsNexradDownloader, RadarCellSegmenter, etc.)
-are NOT required to inherit BaseModule in Step 1. They are wrapped in
-Step 6 of the refactor. BaseModule is the target interface definition.
+Contract functions come from adapt.contracts — import them there, not here.
 """
 
 from abc import ABC, abstractmethod
 from typing import ClassVar
 
-from adapt.contracts import ContractViolation, require  # noqa: F401 — re-exported for callers
-
-# ────────────────────────────────────────────────────────────────────────────
-# BaseModule Interface
-# ────────────────────────────────────────────────────────────────────────────
-
 
 class BaseModule(ABC):
-    """Abstract base for all Adapt processing modules.
+    """Abstract base for all Adapt execution nodes.
 
     Subclasses declare:
     - ``name``: unique identifier used in the execution graph
-    - ``inputs``: list of data keys this module reads from context
-    - ``outputs``: list of data keys this module writes to context
-    - ``input_contracts``: optional {key: callable} validators run before run()
-    - ``output_contracts``: optional {key: callable} validators run after run()
+    - ``inputs``: list of context keys this node reads
+    - ``outputs``: list of context keys this node writes
+    - ``input_contracts``: optional {key: check_fn} validated before run()
+    - ``output_contracts``: optional {key: check_fn} validated after run()
 
-    The graph engine matches ``outputs`` of upstream modules to ``inputs``
-    of downstream modules to resolve execution order automatically.
-    Contract callables are invoked by GraphExecutor automatically — modules
-    do not need to call them manually.
+    The graph engine matches ``outputs`` of upstream nodes to ``inputs``
+    of downstream nodes to resolve execution order automatically.
+    Contract callables (from adapt.contracts) are invoked by GraphExecutor
+    automatically — nodes do not call them manually.
 
     Example::
+
+        from adapt.contracts import check_grid_ds_2d, check_segmented_ds
 
         class DetectModule(BaseModule):
             name = "detection"
             inputs = ["grid_ds_2d"]
             outputs = ["segmented_ds"]
-            input_contracts  = {"grid_ds_2d": assert_gridded}
-            output_contracts = {"segmented_ds": assert_segmented}
+            input_contracts  = {"grid_ds_2d": check_grid_ds_2d}
+            output_contracts = {"segmented_ds": check_segmented_ds}
 
             def run(self, context):
                 grid = context["grid_ds_2d"]
