+++ src/adapt/modules/detection/module.py
@@ -357,66 +357,3 @@ class RadarCellSegmenter:
         old_to_new[labels_sorted] = np.arange(1, len(labels_sorted) + 1)
 
         return old_to_new[labels]
-
-
-# ---------------------------------------------------------------------------
-# BaseModule wrapper — Step 6
-# ---------------------------------------------------------------------------
-
-from adapt.contracts import assert_gridded, assert_segmented  # noqa: E402
-from adapt.execution.module_registry import registry  # noqa: E402
-from adapt.modules.base import BaseModule  # noqa: E402
-
-
-def _check_grid_ds_2d(ds):
-    assert_gridded(ds, "reflectivity")
-
-
-def _check_segmented_ds(ds):
-    assert_segmented(ds, "cell_labels")
-
-
-class DetectModule(BaseModule):
-    """BaseModule wrapper for RadarCellSegmenter.
-
-    Segments convective cells from a 2D reflectivity field using
-    threshold and morphological filtering.
-
-    Context inputs
-    --------------
-    grid_ds_2d : xr.Dataset
-        2D Cartesian dataset (output of LoadModule).
-    config : InternalConfig
-        Runtime configuration.
-
-    Context outputs
-    ---------------
-    segmented_ds : xr.Dataset
-        2D dataset with cell_labels variable added.
-    num_cells : int
-        Number of detected cells.
-    """
-
-    name = "detection"
-    inputs = ["grid_ds_2d", "detection_config"]
-    outputs = ["segmented_ds", "num_cells"]
-    input_contracts  = {"grid_ds_2d": _check_grid_ds_2d}
-    output_contracts = {"segmented_ds": _check_segmented_ds}
-
-    def __init__(self) -> None:
-        self._segmenter = None
-
-    def run(self, context: dict) -> dict:
-        config = context["detection_config"]
-        ds_2d = context["grid_ds_2d"]
-
-        if self._segmenter is None:
-            self._segmenter = RadarCellSegmenter(config)
-
-        segmented = self._segmenter.segment(ds_2d)
-        num_cells = int(segmented[config.labels_var].max().item())
-
-        return {"segmented_ds": segmented, "num_cells": num_cells}
-
-
-registry.register(DetectModule)
